refactor(seascape): extract SiteTerrainPane from the site seascape page
